seperated collabEditing logic in actions
