+++ src/actions/actionTypes.ts
@@ -45,14 +45,9 @@ export const ENSURE_LIBRARIES_LOADED = 'ENSURE_LIBRARIES_LOADED';
 export const EVAL_EDITOR = 'EVAL_EDITOR';
 export const EVAL_REPL = 'EVAL_REPL';
 export const EVAL_TESTCASE = 'EVAL_TESTCASE';
-export const INIT_INVITE = 'INIT_INVITE';
-export const INVALID_EDITOR_SESSION_ID = 'INVALID_EDITOR_SESSION_ID';
-export const FINISH_INVITE = 'FINISH_INVITE';
 export const PLAYGROUND_EXTERNAL_SELECT = 'PLAYGROUND_EXTERNAL_SELECT ';
 export const RESET_WORKSPACE = 'RESET_WORKSPACE';
 export const SEND_REPL_INPUT_TO_OUTPUT = 'SEND_REPL_INPUT_TO_OUTPUT';
-export const SET_EDITOR_SESSION_ID = 'SET_EDITOR_SESSION_ID';
-export const SET_WEBSOCKET_STATUS = 'SET_WEBSOCKET_STATUS';
 export const TOGGLE_EDITOR_AUTORUN = 'TOGGLE_EDITOR_AUTORUN';
 export const UPDATE_CURRENT_ASSESSMENT_ID = 'UPDATE_CURRENT_ASSESSMENT_ID';
 export const UPDATE_CURRENT_SUBMISSION_ID = 'UPDATE_CURRENT_SUBMISSION_ID';
@@ -61,6 +56,13 @@ export const UPDATE_EDITOR_BREAKPOINTS = 'UPDATE_EDITOR_BREAKPOINTS';
 export const UPDATE_HAS_UNSAVED_CHANGES = 'UPDATE_HAS_UNSAVED_CHANGES';
 export const UPDATE_REPL_VALUE = 'UPDATE_REPL_VALUE';
 export const UPDATE_WORKSPACE = 'UPDATE_WORKSPACE';
+
+/** Collab Editing */
+export const INIT_INVITE = 'INIT_INVITE';
+export const INVALID_EDITOR_SESSION_ID = 'INVALID_EDITOR_SESSION_ID';
+export const FINISH_INVITE = 'FINISH_INVITE';
+export const SET_EDITOR_SESSION_ID = 'SET_EDITOR_SESSION_ID';
+export const SET_WEBSOCKET_STATUS = 'SET_WEBSOCKET_STATUS';
 
 /** Session */
 export const FETCH_ANNOUNCEMENTS = 'FETCH_ANNOUNCEMENTS';
