+++ src/actions/workspaces.ts
@@ -174,15 +174,6 @@ export const evalTestcase = (workspaceLocation: WorkspaceLocation, testcaseId: n
   payload: { workspaceLocation, testcaseId }
 });
 
-export const invalidEditorSessionId = () => ({
-  type: actionTypes.INVALID_EDITOR_SESSION_ID
-});
-
-export const finishInvite = (workspaceLocation: WorkspaceLocation) => ({
-  type: actionTypes.FINISH_INVITE,
-  payload: { workspaceLocation }
-});
-
 export const updateEditorValue: ActionCreator<actionTypes.IAction> = (
   newEditorValue: string,
   workspaceLocation: WorkspaceLocation
@@ -227,17 +218,6 @@ export const sendReplInputToOutput: ActionCreator<actionTypes.IAction> = (
   }
 });
 
-export const initInvite: ActionCreator<actionTypes.IAction> = (
-  editorValue: string,
-  workspaceLocation: WorkspaceLocation
-) => ({
-  type: actionTypes.INIT_INVITE,
-  payload: {
-    editorValue,
-    workspaceLocation
-  }
-});
-
 /**
  * Resets a workspace to its default properties.
  *
@@ -266,34 +246,6 @@ export const updateWorkspace = (
   payload: {
     workspaceLocation,
     workspaceOptions
-  }
-});
-
-export const setEditorSessionId: ActionCreator<actionTypes.IAction> = (
-  workspaceLocation: WorkspaceLocation,
-  editorSessionId: string
-) => ({
-  type: actionTypes.SET_EDITOR_SESSION_ID,
-  payload: {
-    workspaceLocation,
-    editorSessionId
-  }
-});
-
-/**
- * Sets sharedb websocket status.
- *
- * @param workspaceLocation the workspace to be reset
- * @param websocketStatus 0: CLOSED 1: OPEN
- */
-export const setWebsocketStatus: ActionCreator<actionTypes.IAction> = (
-  workspaceLocation: WorkspaceLocations,
-  websocketStatus: number
-) => ({
-  type: actionTypes.SET_WEBSOCKET_STATUS,
-  payload: {
-    workspaceLocation,
-    websocketStatus
   }
 });
